backend/requirements.txt
# Core Django
django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1

# LangChain 0.2.x Ecosystem (Stable)
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.40
langchain-google-genai==1.0.10
langchain-openai==0.1.25
langchain-text-splitters==0.2.4

# RAG / Vector Store
chromadb==0.5.18
tiktoken==0.8.0